chore: add utility for resolving import module of an identifier
This utility is necessary in order to write a the `no-classic-classes` rule, which wants to check the module that an identifier comes from against a pattern, rather than a specific value.

The new utility is more flexible than the existing `isIdentifierImportedFrom` helper, and thus can replace it. Rules that used the latter helper function have been updated to use the former instead.